requirements_3.8_ubuntu20.txt
boto3==1.28.84
chromadb==0.4.17
cloudinary==1.38.0
deepface==0.0.89
tensorflow==2.12.0
keras==2.12.0
huggingface-hub==0.20.3
montydb==2.5.2
opencv-python==4.8.1.78
pandas==2.0.3
Pillow==10.1.0
pinecone-client==3.0.2
pymongo==4.6.0
scikit-learn==1.3.2
scipy==1.10.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.2
transformers==4.40.1
pytest-xdist==3.5.0
matplotlib==3.7.5
xgboost==2.0.3
scikit-image==0.21.0
aiohttp==3.9.5
motor==3.4.0
pytest-asyncio==0.23.7